% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tablets}
\alias{tablets}
\title{Tablets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 647 rows and 12 columns.
}
\source{
\url{https://github.com/openrepair/data/tree/master/quests/tablets}
}
\usage{
tablets
}
\description{
Result of an online quest where volunteer participants looked at tablet repair attempts recorded at community events.
}
\details{
Dataset of tablet repair records from the Open Repair Alliance in ORDS format style with added \code{fault_type} values.
Results of \href{https://openrepair.org/open-data/insights/tablets/}{Open Repair Data quest TabiCat"}.

\describe{
\item{id_ords}{ID. Unique identifier from the partner organisation.}
\item{data_provider}{Data provider. Option from ORDS codelist. Name of partner organisation.}
\item{country}{Country. Three letter
\href{https://github.com/openrepair/data/blob/master/tools/iso_country_codes/iso_country_codes.csv}{ISO code},
e.g. “GBR”.}
\item{fault_type_id}{Fault type ID.}
\item{fault_type}{Fault type.
Option from \href{https://github.com/openrepair/data/blob/master/quests/tablets/tabicat_fault_types.csv}{tablet codelist}.}
\item{product_category}{Product category. Option from ORDS
\href{https://standard.openrepair.org/standard.html#ords-product-category-values}{product category codelist}}
\item{brand}{Brand.}
\item{repair_status}{Repair status option from the tablet codelist.}
\item{year_of_manufacture}{Year of manufacture.}
\item{event_date}{Event date. The date of the repair event that the repair took place at.}
\item{language}{Language of problem description. 2 char ISO language code.}
\item{problem}{Problem description.}
}
}
\keyword{datasets}
